requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
fastapi-mail
python-multipart